Fix infinite ice slide and Drag move queue
An Entity would continue attempting to slide when the same Tile was moved to, and would follow in the previous move direction of none. Also, because the ActionQueue clear call, in response to a Drag hit, was moved, Drag could queue behind an ice slide. Drag takes priority over this now by clearing an ice slide queued on the previous frame.